refactor: extract key_generator
