+++ test/e2e/testnet/key_generator.go
@@ -1,0 +1,37 @@
+package testnet
+
+import (
+	"io"
+	"math/rand"
+
+	"github.com/tendermint/tendermint/crypto"
+	"github.com/tendermint/tendermint/crypto/ed25519"
+	"github.com/tendermint/tendermint/crypto/secp256k1"
+)
+
+type keyGenerator struct {
+	random *rand.Rand
+}
+
+func newKeyGenerator(seed int64) *keyGenerator {
+	return &keyGenerator{
+		random: rand.New(rand.NewSource(seed)), //nolint:gosec
+	}
+}
+
+func (g *keyGenerator) Generate(keyType string) crypto.PrivKey {
+	seed := make([]byte, ed25519.SeedSize)
+
+	_, err := io.ReadFull(g.random, seed)
+	if err != nil {
+		panic(err) // this shouldn't happen
+	}
+	switch keyType {
+	case "secp256k1":
+		return secp256k1.GenPrivKeySecp256k1(seed)
+	case "", "ed25519":
+		return ed25519.GenPrivKeyFromSecret(seed)
+	default:
+		panic("KeyType not supported") // should not make it this far
+	}
+}
